2023/11/TestAnAngularInterceptor/src/app/http-403.interceptor.ts
import { Injectable } from '@angular/core';
import {
  HttpRequest,
  HttpHandler,
  HttpEvent,
  HttpInterceptor, HttpErrorResponse
} from '@angular/common/http';
import {catchError, Observable, throwError} from 'rxjs';

@Injectable()
export class Http403Interceptor implements HttpInterceptor {

  errorOccurred = false;
  constructor() {}

  intercept(request: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>> {
    return next.handle(request).pipe(
      catchError((err: HttpErrorResponse) => {
        if (err.status === 403) {
          // do something
          this.errorOccurred = true;
        }
        return throwError(() => err);
      })
    );
  }
}
